Refactoring / feedback from Patrick - updated project references / dependencies - moved UI logic to correct implementation
