+++ SampleProvider/TerminologyProvider/MyTerminologyProvider.cs
@@ -8,108 +8,95 @@ using SampleProvider.Views;
 
 namespace SampleProvider
 {
-    public sealed class MyTerminologyProvider : AbstractTerminologyProvider
-    {
-        private IDefinition _definition;
-        private string _name;
-        private static bool _initialized;
-
-        public override string Name => _name;
-        public override string Description => "Sample terminology provider";
-        public override Uri Uri { get; }
-        public override IDefinition Definition => _definition;
-
-        public MyTerminologyProvider(Uri uri)
-        {
-            Uri = uri;
-            _definition = new Definition(new List<IDescriptiveField>(), new List<IDefinitionLanguage>());
-            _name = $"SAMPLE (SOME-ID/SERVER)";
-            TryLogin();
-            Status = new TerminologyProviderStatus(true);
-        }
-
-        private void TryLogin()
-        {
-            if (_initialized) return;
-            //if the data is not initialized we open a (blocking) window in which the user can login (like Remote MT termbase if not available), then on window close continue init
-            var successfulLogin = ShowConfiguration();
-            _initialized = successfulLogin;
-        }
-
-        private bool ShowConfiguration()
-        {
-            var window = new SampleConfig();
-            ElementHost.EnableModelessKeyboardInterop(window);
-            window.ShowDialog();
-            return true;
-        }
+	public sealed class MyTerminologyProvider : AbstractTerminologyProvider
+	{
+		private IDefinition _definition;
+		private string _name;
 
 
+		public override string Name => _name;
+		public override string Description => "Sample terminology provider";
+		public override Uri Uri { get; }
+		public override IDefinition Definition => _definition;
+
+		public MyTerminologyProvider(Uri uri)
+		{
+			Uri = uri;
+			_definition = new Definition(new List<IDescriptiveField>(), new List<IDefinitionLanguage>());
+			_name = $"SAMPLE (SOME-ID/SERVER)";
+
+			// TODO: this can't work here; the UI is only applicable from the WinFormsUI implementation.
+			// Recommend you move settings, including login to your ITerminologyProviderWinFormsUI implementation
+			//TryLogin();
 
 
-        public override IEntry GetEntry(int id)
-        {
-            return entries.FirstOrDefault(x => x.Id == id);
-        }
+			// TODO: this type of logic should be injected from your factory and WinFormsUI implementations
+			//Status = new TerminologyProviderStatus(true);
+		}
 
-        public override IEntry GetEntry(int id, IEnumerable<ILanguage> languages)
-        {
-            return entries.FirstOrDefault(x => x.Id == id);
-        }
+		public override IEntry GetEntry(int id)
+		{
+			return entries.FirstOrDefault(x => x.Id == id);
+		}
 
-        public override IList<ILanguage> GetLanguages()
-        {
-            return new List<ILanguage>
-            {
-                new EntryLanguage{Locale = new CultureInfo("en-US"), Name = "English"},
-                new EntryLanguage { Locale = new CultureInfo("de-DE"), Name = "German" },
-                new EntryLanguage { Locale = new CultureInfo("fr-FR"), Name = "French" },
-                new EntryLanguage { Locale = new CultureInfo("ja-JA"), Name = "Japanese" },
-            };
-        }
+		public override IEntry GetEntry(int id, IEnumerable<ILanguage> languages)
+		{
+			return entries.FirstOrDefault(x => x.Id == id);
+		}
 
-        private int _idIndex = 1;
-        private List<IEntry> entries = new List<IEntry>();
-        public override IList<ISearchResult> Search(string text, ILanguage source, ILanguage destination, int maxResultsCount, SearchMode mode, bool targetRequired)
-        {
-            var splitted = text.Split();
-            var toReturn = new List<ISearchResult>();
-            foreach (var s in splitted)
-            {
-                var ind = text.IndexOf(s, StringComparison.Ordinal);
-                if (ind < 0) continue;
-                var hit = new SearchMarkupResult
-                {
-                    Language = source,
-                    Id = _idIndex++,
-                    Score = 100,
-                    Text = s,
-                    Positions = new List<IMarkupPosition> { new MarkupPosition { Start = 0, Length = 5 } } //{ Start = ind, Length = s.Length }
-                };
-                var e = new Entry
-                {
-                    Id = hit.Id, 
-                    Fields = new List<IEntryField>(), 
-                    Languages = new List<IEntryLanguage>
-                    {
-                        new EntryLanguage{Locale = source.Locale, Name = source.Name, Fields = new List<IEntryField>(), Terms = new List<IEntryTerm>{ new EntryTerm{Value = s, Transactions = new List<IEntryTransaction>() } }},
-                        new EntryLanguage{Locale = destination.Locale, Name = destination.Name, Fields = new List<IEntryField>(), Terms = new List<IEntryTerm>{ new EntryTerm{Value = s + " TARGET", Transactions = new List<IEntryTransaction>() } }},
-                    },
-                    Transactions = new List<IEntryTransaction>(),
-                };
-                foreach (var language in e.Languages.OfType<EntryLanguage>())
-                {
-                    language.ParentEntry = e;
-                    foreach (var term in language.Terms.OfType<EntryTerm>())
-                    {
-                        term.ParentLanguage = language;
-                    }
-                }
-                entries.Add(e);
-                toReturn.Add(hit);
-            }
-            return toReturn;
-        }
+		public override IList<ILanguage> GetLanguages()
+		{
+			return new List<ILanguage>
+			{
+				new EntryLanguage{Locale = new CultureInfo("en-US"), Name = "English"},
+				new EntryLanguage { Locale = new CultureInfo("de-DE"), Name = "German" },
+				new EntryLanguage { Locale = new CultureInfo("fr-FR"), Name = "French" },
+				new EntryLanguage { Locale = new CultureInfo("ja-JA"), Name = "Japanese" },
+			};
+		}
 
-    }
+		private int _idIndex = 1;
+		private List<IEntry> entries = new List<IEntry>();
+		public override IList<ISearchResult> Search(string text, ILanguage source, ILanguage destination, int maxResultsCount, SearchMode mode, bool targetRequired)
+		{
+			var splitted = text.Split();
+			var toReturn = new List<ISearchResult>();
+			foreach (var s in splitted)
+			{
+				var ind = text.IndexOf(s, StringComparison.Ordinal);
+				if (ind < 0) continue;
+				var hit = new SearchMarkupResult
+				{
+					Language = source,
+					Id = _idIndex++,
+					Score = 100,
+					Text = s,
+					Positions = new List<IMarkupPosition> { new MarkupPosition { Start = 0, Length = 5 } } //{ Start = ind, Length = s.Length }
+				};
+				var e = new Entry
+				{
+					Id = hit.Id,
+					Fields = new List<IEntryField>(),
+					Languages = new List<IEntryLanguage>
+					{
+						new EntryLanguage{Locale = source.Locale, Name = source.Name, Fields = new List<IEntryField>(), Terms = new List<IEntryTerm>{ new EntryTerm{Value = s, Transactions = new List<IEntryTransaction>() } }},
+						new EntryLanguage{Locale = destination.Locale, Name = destination.Name, Fields = new List<IEntryField>(), Terms = new List<IEntryTerm>{ new EntryTerm{Value = s + " TARGET", Transactions = new List<IEntryTransaction>() } }},
+					},
+					Transactions = new List<IEntryTransaction>(),
+				};
+				foreach (var language in e.Languages.OfType<EntryLanguage>())
+				{
+					language.ParentEntry = e;
+					foreach (var term in language.Terms.OfType<EntryTerm>())
+					{
+						term.ParentLanguage = language;
+					}
+				}
+				entries.Add(e);
+				toReturn.Add(hit);
+			}
+			return toReturn;
+		}
+
+	}
 }
