+++ SampleProvider/TerminologyProvider/MyTerminologyProviderFactory.cs
@@ -8,14 +8,18 @@ namespace SampleProvider
                                 Description = "VeryCustom_Terminology_Provider_Description")]
     public class MyTerminologyProviderFactory : ITerminologyProviderFactory
     {
-        public ITerminologyProvider CreateTerminologyProvider(Uri terminologyProviderUri, ITerminologyProviderCredentialStore credentials)
-        {
-            return new MyTerminologyProvider(terminologyProviderUri);
-        }
+	    public ITerminologyProvider CreateTerminologyProvider(Uri terminologyProviderUri, ITerminologyProviderCredentialStore credentials)
+	    {
+		    // TODO Check that your credentials or other criteria are still valid from the criteria persisted in the Uri
+		    // If not valid then throw an exception here (e.g. failed login etc...)
 
-        public bool SupportsTerminologyProviderUri(Uri terminologyProviderUri)
-        {
-            return terminologyProviderUri.AbsoluteUri.StartsWith("very.sample");
-        }
+
+		    return new MyTerminologyProvider(terminologyProviderUri);
+	    }
+
+	    public bool SupportsTerminologyProviderUri(Uri terminologyProviderUri)
+	    {
+		    return terminologyProviderUri.AbsoluteUri.StartsWith("very.sample");
+	    }
     }
 }
