+++ SampleProvider2022/TerminologyProvider/MyTerminologyProviderWinFormsUI.cs
@@ -2,13 +2,17 @@
 using System;
 using System.Linq;
 using System.Windows.Forms;
+using System.Windows.Forms.Integration;
+using SampleProvider.Views;
 
 namespace SampleProvider
 {
     [TerminologyProviderWinFormsUI]
     class MyTerminologyProviderWinFormsUI : ITerminologyProviderWinFormsUI
     {
-        public bool SupportsEditing => true;
+	    private static bool _initialized;
+
+		public bool SupportsEditing => true;
 
         public string TypeDescription => "Terminology Provider for Very Sample Web Services";
 
@@ -16,14 +20,20 @@ namespace SampleProvider
 
         public ITerminologyProvider[] Browse(IWin32Window owner, ITerminologyProviderCredentialStore credentialStore)
         {
-            var uri = new Uri($"very.sample://SOME-URL#####SOME-NAME####SUB-NAME");
+            // TODO initial settings/request for credentials when user is adding your custom provider
+	        //TryLogin();
+
+			var uri = new Uri($"very.sample://SOME-URL#####SOME-NAME####SUB-NAME");
             var provider = new MyTerminologyProvider(uri);
             return new ITerminologyProvider[] { provider };
         }
 
         public bool Edit(IWin32Window owner, ITerminologyProvider terminologyProvider)
         {
-            return false;
+            // TODO edit existing settings/credentials
+	        //TryLogin();
+			
+	        return false;
         }
 
         public TerminologyProviderDisplayInfo GetDisplayInfo(Uri terminologyProviderUri)
@@ -38,5 +48,22 @@ namespace SampleProvider
         {
             return terminologyProviderUri.AbsoluteUri.StartsWith("very.sample");
         }
-    }
+
+
+        private void TryLogin()
+        {
+	        if (_initialized) return;
+	        //if the data is not initialized we open a (blocking) window in which the user can login (like Remote MT termbase if not available), then on window close continue init
+	        var successfulLogin = ShowConfiguration();
+	        _initialized = successfulLogin;
+        }
+
+        private bool ShowConfiguration()
+        {
+	        var window = new SampleConfig();
+	        ElementHost.EnableModelessKeyboardInterop(window);
+	        window.ShowDialog();
+	        return true;
+        }
+	}
 }
